Subforum authentication
I added so the not logged in user cannot see pages of the app besides the register and login.
Subforum management page know who is logged in and when user create the subforum the subforum is created with id of the logged in user. Moderator is needed for restricting the view of the management of the subforums